feat: updated targets column & celery worker

+++ schemas/analytics_schema.py
@@ -1,4 +1,4 @@
 from pydantic import BaseModel
 
 class SetTargetPayload(BaseModel):
-    target: int
+    target: float
